Separate route and controller
